% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggslope.R
\name{ggslope}
\alias{ggslope}
\title{Make a slope chart.}
\usage{
ggslope(
  data,
  times,
  outcome,
  group,
  color = NULL,
  alpha = NULL,
  linetype = NULL,
  size = NULL
)
}
\arguments{
\item{data}{Data}

\item{times}{Times}

\item{outcome}{Outcome}

\item{group}{Group}

\item{color}{Color}

\item{alpha}{Alpha}

\item{linetype}{Linetype}

\item{size}{Size}
}
\value{
A ggplot2 object.
}
\description{
Make a slope chart.
}
