% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggslope.R
\name{ggslope}
\alias{ggslope}
\title{Make a slope chart.}
\usage{
ggslope(
  data,
  times,
  outcome,
  group,
  color = NULL,
  alpha = NULL,
  linetype = NULL,
  size = NULL
)
}
\arguments{
\item{data}{Data}

\item{times}{Times}

\item{outcome}{Outcome}

\item{group}{Group}

\item{color}{Color}

\item{alpha}{Alpha}

\item{linetype}{Linetype}

\item{size}{Size}
}
\value{
A ggplot2 object.
}
\description{
Make a slope chart.
}
\examples{
df <- tidyr::tribble(
~times, ~country,  ~gdp, ~inc,
"1990",   "A",  22.3, TRUE,
"2000",   "A",  44.6, TRUE,
"1990",   "B",  12.3, FALSE,
"2000",   "B",  4.6, FALSE
)
ggslope(df,
        times = times,
        outcome = gdp,
        group = country)
}
